Fix touch event handler - respect zOrder defined in the skin
The old implementation would iterate the elements in ascending zOrder - but we would want the top most widget. Therefore only keep track of the last widget in the stack, and let it handle the event.

This partially renders an earlier commit obsolete, where I disabled the touch event on the artwork to work around this limitation.