Force use of IPv4 UDP socket
The commit to resolve:
  https://github.com/diegonehab/luasocket/issues/147
in lua-socket to support IPv6 has changed the behaviour of socket.udp()
so that the socket isn't created until the first packet is transmitted.
This is to allow the create family to be used. But this stops us from
being up to set the broadcast flag on the socket, so no socket exists.
The result in jivelite is that we fail to discover the SlimServer.

This commit looks for the existence of the udp4 function and calls that.
This allows us to support current lua-socket, and older deployments.

In an ideal world we'd support IPv6.